src/scraper/requirements.txt
pandas>=3.0.1
PyYAML>=6.0.2
selenium>=4.41.0
webdriver_manager>=4.0.2
